+++ src/main/java/dev/louis/nebula/spell/source/BlockEntitySpellSource.java
@@ -64,11 +64,11 @@ public class BlockEntitySpellSource<BE extends BlockEntity> implements SpellSour
 
     @Override
     public void drainMana(long amount, TransactionContext context) throws SpellFumble {
-        getManaPool().map(manaPool -> (manaPool.extractMana(amount, context) == amount)).filter(Boolean::booleanValue).orElseThrow(SpellFumble::new);
+        getManaPool().map(manaPool -> (manaPool.extractMana(amount, context) == amount)).filter(Boolean::booleanValue).orElseThrow(SpellFumble::manaFumble);
     }
 
     @Override
     public void startSpellEffect(SpellEffect spellEffect, TransactionContext transaction) throws SpellFumble {
-        throw new SpellFumble();
+        throw SpellFumble.spellEffectFumble();
     }
 }